Include rule workspace in load paths
I've found that if you have a sass_binary in an external workspace/module, then file references inside that workspace will not work correctly. This is because sources and outputs from the rule end up in a subfolder of `external` rather than the execroot, bindir or gendir directly.

This change loads the root of the workspace from the respective folders iff the rule is built in an external workspace.

I added a test to verify this fix works. Removing the changes to `defs.bzl` will fail the smoke test.